% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_endpoint_functions.R
\name{get_daily_geojson}
\alias{get_daily_geojson}
\title{Daily Waldbrandgefahrenindex geojson}
\usage{
get_daily_geojson()
}
\value{
a geojson
}
\description{
Daily Waldbrandgefahrenindex geojson
}
